+++ lib/sorcery/test_helpers/rails/controller.rb
@@ -1,0 +1,17 @@
+module Sorcery
+  module TestHelpers
+    module Rails
+      module Controller 
+        def login_user(user = nil, test_context = {})
+          user ||= @user
+          @controller.send(:auto_login, user)
+          @controller.send(:after_login!, user, [user.send(user.sorcery_config.username_attribute_names.first), 'secret'])
+        end
+
+        def logout_user
+          @controller.send(:logout)
+        end
+      end
+    end
+  end
+end
